research_crew/tools/pdf_extractor.py
"""
pdf_extractor.py
----------------
Fetches a PDF from a URL and extracts clean plain text using PyMuPDF (fitz).

Safety limits
-------------
* Fetched content is capped at 10 MB to avoid memory issues.
* Extracted text is truncated to HARD_TEXT_LIMIT (3 000 chars).
* Only the first MAX_PAGES pages are processed.
"""

import io
import logging
from typing import Type

import requests
from pydantic import BaseModel, Field
from tenacity import retry, retry_if_exception_type, stop_after_attempt, wait_exponential

try:
    from crewai.tools import BaseTool
except ImportError:
    from langchain.tools import BaseTool  # type: ignore

from research_crew.utils.token_utils import truncate_text

logger = logging.getLogger(__name__)

HARD_TEXT_LIMIT = 3000
MAX_PAGES = 15
MAX_DOWNLOAD_MB = 10


class _PDFInput(BaseModel):
    url: str = Field(..., description="Direct URL to a PDF file")


class PDFExtractorTool(BaseTool):
    name: str = "PDF Text Extractor"
    description: str = (
        "Download a PDF from a URL and extract readable plain text. "
        "Returns at most 3000 characters."
    )
    args_schema: Type[BaseModel] = _PDFInput

    @retry(
        retry=retry_if_exception_type(requests.exceptions.RequestException),
        wait=wait_exponential(multiplier=2, min=2, max=20),
        stop=stop_after_attempt(3),
        reraise=True,
    )
    def _run(self, url: str) -> str:
        try:
            import fitz  # PyMuPDF
        except ImportError:
            return "Error: PyMuPDF not installed. Run: pip install PyMuPDF"

        # 🔧 Fix arXiv links
        if "arxiv.org/abs/" in url:
            url = url.replace("/abs/", "/pdf/")
            if not url.endswith(".pdf"):
                url += ".pdf"

        logger.info("Downloading PDF: %s", url)

        resp = requests.get(
            url,
            headers={"User-Agent": "ResearchCrew/1.0"},
            timeout=30,
            stream=True,
        )
        resp.raise_for_status()

        # 🔴 Validate content type
        if "application/pdf" not in resp.headers.get("Content-Type", ""):
            return "Error: URL does not point to a valid PDF file."

        # 📥 Download with size cap
        content = b""
        max_bytes = MAX_DOWNLOAD_MB * 1024 * 1024

        for chunk in resp.iter_content(chunk_size=8192):
            content += chunk
            if len(content) > max_bytes:
                logger.warning("PDF exceeds size limit (%d MB)", MAX_DOWNLOAD_MB)
                break

        # 📄 Parse PDF
        doc = fitz.open(stream=io.BytesIO(content), filetype="pdf")
        pages_read = min(len(doc), MAX_PAGES)

        text_parts = []
        for i in range(pages_read):
            page = doc[i]
            text_parts.append(page.get_text("text"))

        doc.close()

        full_text = "\n".join(text_parts)
        full_text = _clean_pdf_text(full_text)

        # 🔴 Handle empty text
        if not full_text.strip():
            return "Error: No readable text extracted from PDF."

        # ✅ Token-safe truncation
        truncated = truncate_text(full_text, HARD_TEXT_LIMIT)

        logger.info("Extracted %d chars from %d pages", len(truncated), pages_read)

        return f"[PDF CONTENT]\n\n{truncated}"


def _clean_pdf_text(text: str) -> str:
    """Clean PDF artefacts"""
    import re

    text = text.replace("\x00", "").replace("\x0c", "\n")
    text = re.sub(r"[ \t]{2,}", " ", text)

    lines = [ln.strip() for ln in text.splitlines()]
    lines = [ln for ln in lines if ln]

    return "\n".join(lines)